requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
Django==5.0.2
djangorestframework==3.14.0
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.6
multidict==6.0.5
mysqlclient==2.2.4
packaging==23.2
pytz==2024.1
sqlparse==0.4.4
yarl==1.9.4
